requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
python-docx>=0.8.11
gspread>=5.11.0
oauth2client>=4.1.3
openpyxl>=3.1.0
